\name{fastlpr_plot}
\alias{fastlpr_plot}
\title{Plot regression results}
\usage{
fastlpr_plot(fpp_yhat, x_range = NULL, n_points = 1000, ...)
}
\description{
Visualizes fitted regression results.
}
\arguments{
  \item{fpp_yhat}{Interpolant function or result object.}
  \item{x_range}{Optional range for plotting.}
  \item{n_points}{Number of points for plotting (default: 1000).}
  \item{...}{Additional plotting parameters.}
}
\seealso{
\code{\link{cv_fastlpr}}
}
\keyword{hplot}
