\name{is_fastlpr_result}
\alias{is_fastlpr_result}
\title{Check if Object is a fastlpr_result}
\description{
Tests whether an object is of class \code{fastlpr_result}.
}
\usage{
is_fastlpr_result(x)
}
\arguments{
  \item{x}{An R object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{"fastlpr_result"},
\code{FALSE} otherwise.
}
\examples{
is_fastlpr_result(list())  # FALSE
}
