\name{set_defaults}
\alias{set_defaults}
\title{Set default options}
\usage{
set_defaults(opt, field, default)
}
\description{
Utility function to set a single default option field.
}
\arguments{
  \item{opt}{User-provided options list.}
  \item{field}{Field name to set.}
  \item{default}{Default value for the field.}
}
\value{
Merged options list.
}
\keyword{utilities}
